fix: move all the providers to separate component
