+++ src/App.js
@@ -5,27 +5,18 @@ import Dashboard from "./pages/dashboard/Dashboard";
 import {Route, Routes, Navigate} from "react-router-dom";
 import DashboardRoutes from "./pages/dashboard/Routes";
 import DashboardRequests from "./pages/dashboard/Requests";
-import {createTheme, ThemeProvider} from "@mui/material";
 
-const theme = createTheme({
-  typography: {
-    fontFamily: ["Nunito", "sans-serif"].join(","),
-    fontWeightMedium: "bold",
-  },
-});
 function App() {
   return (
     <div>
-      <ThemeProvider theme={theme}>
-        <Routes>
-          <Route path="/" element={<Login />} />
-          <Route path="/dashboard" element={<Dashboard />}>
-            <Route index element={<Navigate to="/dashboard/routes" />} />
-            <Route path="requests" element={<DashboardRequests />} />
-            <Route path="routes" element={<DashboardRoutes />} />
-          </Route>
-        </Routes>
-      </ThemeProvider>
+      <Routes>
+        <Route path="/" element={<Login />} />
+        <Route path="/dashboard" element={<Dashboard />}>
+          <Route index element={<Navigate to="/dashboard/routes" />} />
+          <Route path="requests" element={<DashboardRequests />} />
+          <Route path="routes" element={<DashboardRoutes />} />
+        </Route>
+      </Routes>
     </div>
   );
 }
